Client: Move environment to src directory.
