+++ client/src/app/core/services/logger.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { LogLevel } from '../../utils/constants';
 
 @Injectable({
