+++ client/src/app/core/services/websocket-connection.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { LoggerService } from './logger.service';
 import { Router } from '@angular/router';
