+++ client/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  apiUrl: 'https://api.mydomain.com',
+  apiUrl: 'api.mydomain.com',
 };
