+++ client/src/main.server.ts
@@ -1,5 +1,5 @@
 import { enableProdMode } from '@angular/core';
-import { environment } from './app/environments/environment';
+import { environment } from './environments/environment';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
 import { config } from './app/app.config.server';
